+++ modules/connectors/landmark/karrio/mappers/landmark/proxy.py
@@ -20,7 +20,6 @@ class Proxy(rating_proxy.RatingMixinProxy, proxy.Proxy):
             method="POST",
             headers={
                 "Content-Type": "application/xml",
-                "User-Agent": "Karrio/2024.0 (https://karrio.io)",
             },
         )
 
@@ -34,7 +33,6 @@ class Proxy(rating_proxy.RatingMixinProxy, proxy.Proxy):
             method="POST",
             headers={
                 "Content-Type": "application/xml",
-                "User-Agent": "Karrio/2024.0 (https://karrio.io)",
             },
         )
 
@@ -52,7 +50,6 @@ class Proxy(rating_proxy.RatingMixinProxy, proxy.Proxy):
                     headers={
                         "Content-Type": "application/xml",
                         "Cache-Control": "no-cache",
-                        "User-Agent": "Karrio/2024.0 (https://karrio.io)",
                     },
                 ),
             ),